requirements.txt
Flask
reportlab
pywebview
pyinstaller
